requirements.txt
Flask
pychromecast
unicode_slugify==0.1.3
pathlib==1.0.1
